refactor: fold qk_norm_rope_quant_gfx1250 into qk_norm_rope_quant
The two modules implement the same public contract; gfx1250.py was made by
copying the wave64 file and swapping the codegen. Keeping them apart made the
duplication a correctness risk rather than just untidiness: the wave64 entry
point delegates to gfx1250 *before* running its own validation, so each copy
only ever executes on hardware where the other is dead, and a validation fix
in one silently misses the other.

This is the mechanical half -- pure code motion, no logic change:

  - both builders kept verbatim as _build_kernel_w64 / _build_kernel_w32, each
    opening with a local BLOCK_THREADS rebind off the new module-level
    BLOCK_THREADS_W64 / BLOCK_THREADS_W32 (the old per-module BLOCK_THREADS
    was the only real name collision; it is referenced solely inside the
    builders, so the rebind mirrors the existing `ROWS_PER_WG = rows_per_wg`
    pattern);
  - _store_fp8_packed disambiguated to _w64 / _w32; the other store helpers
    already had distinct names;
  - one copy each of _SQRT2, _fp8_const, SCALE_DTYPE_*, _TORCH_DTYPE_FOR_SCALE
    and _DEFAULT_COMPILE_HINTS (they were byte-identical apart from the
    _fp8_const docstring, where the fuller gfx1250 wording survives);
  - the arch dispatch now calls the local function instead of importing a
    sibling module.

Also drops GROUP_SIZE_OPTIONS, which had no references anywhere in the repo.
It has no leading underscore, so flagging it here rather than in passing.

Host-side dedup and the _build_kernel merge follow separately.

Verification -- wave64 cannot be executed on this box, so codegen equivalence
is established by dumping FlyDSL's arch-independent MLIR stages (00-09;
permlane32 first appears at stage 10) across a config matrix before and after:

  - wave32: 44 configs, 484 artifacts including all 44 21_final_isa.s
    -- byte-identical;
  - wave64: 22 configs, 220 artifacts, reached by patching get_gfx so the
    entry point stops delegating -- byte-identical;
  - op_tests/test_flydsl_qk_norm_rope_quant.py: all 234 rows x 2 error
    columns equal to HEAD, including the pre-existing T=65540 H>=64 failure
    (Q output exceeds the 4 GB buffer-descriptor limit);
  - interleaved A/B vs HEAD: decode 3.399us vs 3.422us (0.993x, inside the
    +/-0.2us run-to-run spread); prefill T=16384 435.55us vs 437.63us and
    441.58us vs 443.62us with q_weight (0.995x both).
